+++ src/main/java/com/eunsun/travel_mate/repository/jpa/TourInfoRepository.java
@@ -1,6 +1,6 @@
-package com.eunsun.travel_mate.repository;
+package com.eunsun.travel_mate.repository.jpa;
 
-import com.eunsun.travel_mate.domain.TourInfo;
+import com.eunsun.travel_mate.domain.tourInfo.TourInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
